fraud-detection/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
joblib==1.3.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
alembic==1.13.1
# CVE-2024-33663, CVE-2024-33664: python-jose has unfixable algorithm-confusion
# vulnerabilities; replaced with PyJWT which enforces strict algorithm validation.
PyJWT==2.8.0
cryptography>=42.0.5
passlib[bcrypt]==1.7.4
httpx==0.27.0
prometheus-fastapi-instrumentator==6.1.0
# CVE-2024-1135: ReDoS in python-multipart < 0.0.10; fixed in 0.0.12
python-multipart==0.0.12
structlog==24.1.0
